+++ src/cjs/utils/datastructures/Girih.js
@@ -5,6 +5,7 @@
  * @modified 2020-11-25 Ported to TypeScript from vanilla JS.
  * @modified 2024-03-10 Fixed some types for Typescript 5 compatibility.
  * @modified 2026-01-06 Added method `Girih.locateContainingTileAndEdge` to locate tile/edge pairs.
+ * @modified 2026-01-12 Added method `Girih.getTileByCenter` to locate tiles by position.
  * @version  1.1.0
  * @file     Girih
  **/
@@ -98,6 +99,7 @@ var Girih = /** @class */ (function () {
         this.tiles.splice(index, 1);
     };
     /**
+     * Replace all current tiles with the given ones.
      *
      * @param tiles
      */
@@ -106,6 +108,24 @@ var Girih = /** @class */ (function () {
         for (var i in tiles) {
             this.addTile(tiles[i]);
         }
+    };
+    /**
+     * Find the tile with the given center.
+     *
+     * @name getTileByCenter
+     * @memberof Girih
+     * @instance
+     * @param {XYCoords} center - The center point to look for.
+     * @return {GirihTile} The tile or null if not found.
+     */
+    Girih.prototype.getTileByCenter = function (center) {
+        for (var i = 0; i < this.tiles.length; i++) {
+            var pos = this.tiles[i].position;
+            if (pos === center || (pos.x == center.x && pos.y == center.y)) {
+                return this.tiles[i];
+            }
+        }
+        return null;
     };
     /**
      * Find that tile (index) which contains the given position. First match will be returned.