+++ src/esm/utils/datastructures/Girih.js
@@ -4,6 +4,7 @@
  * @modified 2020-11-25 Ported to TypeScript from vanilla JS.
  * @modified 2024-03-10 Fixed some types for Typescript 5 compatibility.
  * @modified 2026-01-06 Added method `Girih.locateContainingTileAndEdge` to locate tile/edge pairs.
+ * @modified 2026-01-12 Added method `Girih.getTileByCenter` to locate tiles by position.
  * @version  1.1.0
  * @file     Girih
  **/
@@ -95,6 +96,7 @@ export class Girih {
         this.tiles.splice(index, 1);
     }
     /**
+     * Replace all current tiles with the given ones.
      *
      * @param tiles
      */
@@ -103,6 +105,24 @@ export class Girih {
         for (var i in tiles) {
             this.addTile(tiles[i]);
         }
+    }
+    /**
+     * Find the tile with the given center.
+     *
+     * @name getTileByCenter
+     * @memberof Girih
+     * @instance
+     * @param {XYCoords} center - The center point to look for.
+     * @return {GirihTile} The tile or null if not found.
+     */
+    getTileByCenter(center) {
+        for (var i = 0; i < this.tiles.length; i++) {
+            const pos = this.tiles[i].position;
+            if (pos === center || (pos.x == center.x && pos.y == center.y)) {
+                return this.tiles[i];
+            }
+        }
+        return null;
     }
     /**
      * Find that tile (index) which contains the given position. First match will be returned.